ft: split aws-sdk tests
Split aws-sdk tests so that parallel CI containers can be leveraged to run
different test suites.
This introduces a manual step when adding a new test directory under aws-sdk as
the end to end CI no longer just runs `npm run ft_test`. For the tests to run in
end to end CI, the new dir name needs to be added in S3/package.json to either
- ft_awssdk_buckets
- ft_awssdk_objects_misc
- ft_awssdk_versioning

If a new npm command is being introduced, make sure to update Integration scripts
to run the test suite.